Fixed Angular injection and form verification
errors.minlength needs to be prefixed with errors?.minlegth to avoid throwing a null value read error. AngularFireDatabase and AngularFireAuth need to be included as providers in app.module.ts. AuthGuard was relying on a property from the authservice that didn't exist, changed to return true if user is authorized.